[MINOR][TESTS] Unify workspace test helpers via SparkTestPaths trait
### What changes were proposed in this pull request?

Introduces `org.apache.spark.util.SparkTestPaths` in `common-utils` test sources, holding the two helpers that are currently duplicated across module boundaries:

- `getWorkspaceFilePath(parts...)` — resolve a path relative to `spark.test.home` / `SPARK_HOME`.
- `regenerateGoldenFiles` — env-var check for `SPARK_GENERATE_GOLDEN_FILES=1`.

The trait is mixed into:
- `SparkTestSuite` (core) — drops local copies of both helpers.
- `ConnectFunSuite` (connect client) — drops the local `getWorkspaceFilePath` (which carried a `// Borrowed from SparkFunSuite` comment).
- `LogKeysSuite` (common-utils) — drops local copies of both.

Also removes the now-redundant `private val regenerateGoldenFiles` in `PlanGenerationTestSuite` (it would have shadowed/clashed with the inherited `protected def`).

Living in `common-utils` is required because some consumers (`LogKeysSuite` in `common-utils`, `ConnectFunSuite` in the shaded Spark Connect client) cannot reach `SparkFunSuite` / `SparkTestSuite` in `spark-core` due to module dependency direction.

### Why are the changes needed?

Cleanup. The two helpers are byte-for-byte identical at three call sites (four counting the `regenerateGoldenFiles` clone), and the connect copy explicitly comments that it was borrowed from `SparkFunSuite`. Consolidating prevents drift.

### Does this PR introduce _any_ user-facing change?

No. Test-only refactor.

### How was this patch tested?

`build/sbt common-utils/Test/compile core/Test/compile connect-client-jvm/Test/compile sql/Test/compile` succeeds. No behavior change.

### Was this patch authored or co-authored using generative AI tooling?

Generated-by: Claude Code (Opus 4.7)
